+++ eslint.config.mjs
@@ -1,5 +1,9 @@
+import { createRequire } from 'node:module'
 import antfu from '@antfu/eslint-config'
 import * as VueVineESLintParser from '@vue-vine/eslint-parser'
+
+const require = createRequire(import.meta.url)
+const VueVineESLintParserPackageJSON = require('@vue-vine/eslint-parser/package.json')
 
 export default antfu(
   {
@@ -25,9 +29,16 @@ export default antfu(
     files: [
       'packages/e2e-test/**/*.vine.ts',
       'packages/playground/**/*.vine.ts',
+      'packages/nuxt-module/playground/**/*.vine.ts',
     ],
     languageOptions: {
-      parser: VueVineESLintParser,
+      parser: {
+        meta: {
+          name: '@vue-vine/eslint-parser',
+          version: VueVineESLintParserPackageJSON.version,
+        },
+        ...VueVineESLintParser,
+      },
     },
     rules: {
       'no-console': 'off',