+++ packages/nuxt-module/test/fixtures/basic/app.vue
@@ -1,0 +1,8 @@
+<script setup lang="ts">
+import { VineTest } from './test.vine'
+</script>
+
+<template>
+  <div>basic</div>
+  <VineTest />
+</template>