+++ scripts/run-test.js
@@ -3,6 +3,7 @@ import { runCommand } from './utils'
 const testCommand = 'cross-env NODE_ENV=test '
   + 'pnpm run test:compiler --run && '
   + 'pnpm run test:e2e --run && '
-  + 'pnpm run test:eslint-parser'
+  + 'pnpm run test:eslint-parser && '
+  + 'pnpm run test:nuxt'
 
 runCommand(testCommand)