+++ packages/vite-plugin/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @vue-vine/vite-plugin
 
+## 0.1.20
+
+### Patch Changes
+
+- fix vite-plugin read styleDefine of undefined.
+
+## 0.1.19
+
+### Patch Changes
+
+- Update version with SSR support features.
+- Fix again HMR style update after support multiple vineStyle.
+
+## 0.1.18
+
+### Patch Changes
+
+- Fix HMR style update after support multiple vineStyle.
+
+## 0.1.16
+
+### Patch Changes
+
+- Fix package exports specification for Node ESM and error capture in Vite plugin.
+
 ## 0.1.1
 
 ### Patch Changes